All methods to run parallel operations on collections moved to collections.scala file
